Add files via upload
Hi, guys. I upload a very simple httpserver just for testing. It is based on TCP. Please run the server and let app to connect to it. Then the app will receive a json file. If it fails, plz tell me as soon as possible.
If you dont have environment for python, plz install Anaconda(python 2.7). It is very convenient.